+++ src/commands/log/list.ts
@@ -2,11 +2,21 @@ import columnify from "columnify";
 import { Session } from "../../entities/session";
 import { dateToTimeString, durationToString } from "../../utils";
 import * as style from "../../style";
+import { EOL } from "os";
 
 export function sessionsToList(sessions: Session[]): string {
-  return columnify(
-    sessions
-      .map(({ id, project, start, end, startSeconds, endSeconds }) => [
+  let output = "";
+  for (const {
+    id,
+    project,
+    start,
+    end,
+    startSeconds,
+    endSeconds,
+    notes,
+  } of sessions) {
+    output += columnify(
+      [
         { label: "Session:", value: style.light(id) },
         { label: "Project:", value: style.project(project.name) },
         project.client
@@ -21,13 +31,23 @@ export function sessionsToList(sessions: Session[]): string {
           label: "Duration:",
           value: style.bold(durationToString(endSeconds - startSeconds)),
         },
-        { label: "", value: "" },
-      ])
-      .flat()
-      .filter((row) => !!row),
-    {
-      showHeaders: false,
-      columnSplitter: "  ",
+      ].filter((row) => !!row),
+      {
+        showHeaders: false,
+        columnSplitter: "  ",
+      }
+    );
+
+    if (notes) {
+      const trimmedNotes = notes.replace(/^\s+|\s+$/g, "");
+      if (trimmedNotes) {
+        output +=
+          EOL + style.light(trimmedNotes.replace(/^|\n/g, "\n  ")) + EOL;
+      }
     }
-  );
+
+    output += EOL;
+  }
+
+  return output;
 }