+++ src/entities/state.ts
@@ -4,6 +4,7 @@ export type StateSession = {
   project: Project;
   start: Date;
   tags?: string[];
+  notes?: string;
 };
 
 export type State = {